# contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: Doublangu API
  version: "v1"
  description: |
    Single-owner language-learning media system. JSON API endpoints return
    structured error envelopes for non-2xx responses; plugin asset file
    responses retain their content-serving contract.
servers:
  - url: http://localhost:8080
    description: Local development server.
paths:
...
  /api/v1/auth/login:
    post:
      summary: Authenticate as the single owner
      description: |
        Requires a CSRF token (X-CSRF-Token header + csrf_token cookie).
        Rate-limited to 5 attempts per minute per IP.
        Sets doublangu_session cookie on success.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  description: Owner password.
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "400":
          description: Invalid JSON request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "429":
          description: Rate limited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "405":
          description: Method not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: Session or CSRF refresh failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /api/v1/auth/csrf:
    get:
      summary: Bootstrap browser CSRF protection
      description: |
        Issues the same-origin csrf_token cookie required by login and logout.
        The token is read by the browser and echoed in X-CSRF-Token.
      responses:
        "200":
          description: CSRF cookie issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OKResponse"
        "405":
          description: Method not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: CSRF bootstrap failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /api/v1/auth/logout:
    post:
      summary: End the current session
      description: |
        Requires the same CSRF cookie/header pair as login. Deletes the session
        from the database and clears the cookie after CSRF validation.
      responses:
        "200":
          description: Logout complete.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OKResponse"
        "403":
          description: Missing or invalid CSRF token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "405":
          description: Method not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: Session deletion failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /api/v1/auth/session:
    get:
      summary: Check session validity
      description: Returns whether the current doublangu_session cookie is valid.
      responses:
        "200":
          description: Session status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
        "405":
          description: Method not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /api/v1/ui/contributions:
    get:
      summary: Read enabled UI plugin contributions
      description: Requires an authenticated owner session.
      responses:
        "200":
          description: Versioned plugin contribution snapshot.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Owner session required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "405":
          description: Method not allowed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: Contribution registry failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"

  # --- Article reader endpoints ---
  # All reader routes require an authenticated owner session. Mutations also
  # require the CSRF double-submit cookie/header pair.
  /api/v1/articles:
    get:
      summary: List saved reader articles
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Article summaries, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/Internal"

    post:
      summary: Save a pasted reader article
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCreate"
      responses:
        "201":
          description: Article saved immediately with background semantic analysis queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/articles/{id}:
    get:
      summary: Get a reader article
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "200":
          description: Article with ordered blocks and annotations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/articles/{id}/enrich:
    post:
      summary: Queue reader analysis (compatibility alias)
      description: Deprecated compatibility alias for reanalyze. Returns immediately with a queued article.
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "202":
          description: Article queued for background analysis.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/articles/{id}/reanalyze:
    post:
      summary: Queue explicit background article reanalysis
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "202":
          description: New analysis work was queued; an older accepted analysis remains readable until replacement.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/articles/{id}/narration:
    get:
      summary: Get the ordered sentence narration manifest
      security: [{ sessionCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "200":
          description: Narration status and ordered sentence clips.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Narration" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      summary: Queue narration regeneration
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "202":
          description: Missing or purged sentence renders were queued.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Article" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/Internal" }
    delete:
      summary: Clear article narration storage
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "200":
          description: Narration bindings cleared and orphaned long-form bytes reclaimed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NarrationClearResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/learning-state:
    put:
      summary: Set an article annotation learning state
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/SemanticLearningStateInput"
                - $ref: "#/components/schemas/LearningStateInput"
      responses:
        "200":
          description: Stored sense-keyed or legacy compatibility learning state.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SemanticLearningState"
                  - $ref: "#/components/schemas/LearningState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/speech-workers/enrollments:
    post:
      summary: Create a one-time speech worker enrollment secret
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses:
        "201":
          description: Secret displayed once to the owner.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkerEnrollment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/speech-workers:
    get:
      summary: List enrolled speech workers
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Worker status without credential hashes.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/SpeechWorker" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
  /api/v1/speech-workers/{id}:
    delete:
      summary: Revoke a speech worker and return its leases to the queue
      security: [{ sessionCookie: [], csrfToken: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
      responses:
        "200": { description: Worker revoked., content: { application/json: { schema: { $ref: "#/components/schemas/OKResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/speech-worker/enroll:
    post:
      summary: Exchange a one-time secret for a worker credential
      parameters:
        - name: X-Doublangu-Enrollment-Token
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerEnrollmentRequest" }
      responses:
        "201":
          description: Worker credential; shown only during enrollment.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkerEnrollmentResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { description: Invalid or expired enrollment secret., content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } } }
  /api/v1/speech-worker/lease:
    post:
      summary: Long-poll and lease one compatible speech job
      parameters:
        - name: X-Doublangu-Worker-Token
          in: header
          required: true
          schema: { type: string, minLength: 32 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerLeaseRequest" }
      responses:
        "200":
          description: One leased render job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkerLease" }
        "204": { description: No compatible work before the long-poll deadline. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { description: Worker credential required., content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } } }
  /api/v1/speech-worker/jobs/{id}/heartbeat:
    post:
      summary: Renew a speech job lease
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ULID" }
        - name: X-Doublangu-Worker-Token
          in: header
          required: true
          schema: { type: string }
        - name: X-Doublangu-Lease-Token
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkerHeartbeat" }
      responses:
        "200": { description: Lease renewed., content: { application/json: { schema: { $ref: "#/components/schemas/WorkerHeartbeatResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { description: Worker credential required., content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } } }
        "409": { description: Lease expired or canceled., content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } } }
...
components:
  schemas:
    APIError:
      type: object
      additionalProperties: false
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Versioned machine-readable error code (for example v1.csrf_error).
    ULID:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      description: Canonical uppercase Crockford ULID.
...
    OKResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean

    # --- Article reader schemas ---
...
    EnrichmentStatus:
      type: string
      enum: [draft, processing, ready, failed]
...
    ArticleSummary:
      type: object
      required: [id, title, source_language, target_language, enrichment_status,
        enrichment_error_code, created_at, updated_at, content_hash, analysis_status,
        analysis_error_code, narration_status, narration_error_code]
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        title: { type: string }
        source_language: { $ref: "#/components/schemas/LanguageTag" }
        target_language: { $ref: "#/components/schemas/LanguageTag" }
        enrichment_status: { $ref: "#/components/schemas/EnrichmentStatus" }
        enrichment_error_code: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
        content_hash: { $ref: "#/components/schemas/SHA256" }
        analysis_status: { $ref: "#/components/schemas/AnalysisStatus" }
        analysis_error_code: { type: string }
        narration_status: { $ref: "#/components/schemas/NarrationStatus" }
        narration_error_code: { type: string }
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    Forbidden:
      description: Missing or invalid CSRF token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
...
    Internal:
      description: Unexpected server failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    MethodNotAllowed:
      description: Method not allowed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }